+++ src/features/prop-types/test/node.spec.js
@@ -1,0 +1,1 @@
+describe('@skypager/prop-types', function() {})